+++ plantcv/geospatial/__init__.py
@@ -5,6 +5,7 @@ from plantcv.geospatial import read
 from plantcv.geospatial import convert
 from plantcv.geospatial import analyze
 from plantcv.geospatial import create_shapes
+from plantcv.geospatial.subtract_dsm import subtract_dsm
 from plantcv.geospatial.center_grid_rois import center_grid_rois
 from plantcv.geospatial.images import Image, GEO, DSM
 
@@ -18,6 +19,7 @@ __all__ = [
     "read",
     "analyze",
     "create_shapes",
+    "subtract_dsm",
     "center_grid_rois",
     "convert",
     "Image",